fix compilation warnings from the future

+++ Sources/VLCGoogleDriveController.m
@@ -191,7 +191,7 @@
         return;
     }
 
-    APLog(@"found filtered metadata for %lu files", _currentFileList.count);
+    APLog(@"found filtered metadata for %lu files", (unsigned long)_currentFileList.count);
     if ([self.delegate respondsToSelector:@selector(mediaListUpdated)])
         [self.delegate mediaListUpdated];
 }
